+++ lib/ReactViews/DataCatalog/CatalogGroup.jsx
@@ -135,14 +135,16 @@ function CatalogGroup(props) {
               <Loader />
             </li>
           )}
-          {!props.loading && props.children.length === 0 && props.emptyMessage && (
-            <li
-              className={classNames(Styles.label, Styles.labelNoResults)}
-              key="empty"
-            >
-              {props.emptyMessage}
-            </li>
-          )}
+          {!props.loading &&
+            props.children.length === 0 &&
+            props.emptyMessage && (
+              <li
+                className={classNames(Styles.label, Styles.labelNoResults)}
+                key="empty"
+              >
+                {props.emptyMessage}
+              </li>
+            )}
 
           {!props.loading ? props.children : null}
         </ul>